UI rotation
Fixed a issue where any additional years would be cut off without being dispalyed on devices that are small and rotated.
The screen will now increase the height if th screen is rotated and a small screen. allowing for the user to scroll down to view the full year if they want.